[Refactoring] Se traslada metodo GetArticulosPorAutor a BuscadorArticuloDBHandler
